commonize awaitSkiko (still internal) and remove getSkikoWasm completely
