--- skiko/src/jsMain/kotlin/org/jetbrains/skia/webext/ArrayBuffer.js.kt
@@ -1,8 +1,0 @@
-package org.jetbrains.skia.webext
-
-import kotlinx.coroutines.await
-import org.jetbrains.skiko.wasm.awaitSkiko
-
-internal actual suspend fun getSkikoWasm(): SkikoWasm {
-    return awaitSkiko.await().unsafeCast<SkikoWasm>()
-}
