+++ skiko/src/jsMain/kotlin/org/jetbrains/skiko/wasm/WrapperExternal.kt
@@ -4,4 +4,4 @@ package org.jetbrains.skiko.wasm
 
 import kotlin.js.Promise
 
-internal external val awaitSkiko: Promise<Any>
+internal actual external val awaitSkiko: Promise<JsAny>
