+++ skiko/src/wasmJsMain/kotlin/org/jetbrains/skiko/wasm/WrapperExternal.kt
@@ -3,4 +3,4 @@ package org.jetbrains.skiko.wasm
 
 import kotlin.js.Promise
 
-internal external val awaitSkiko: Promise<JsAny?>
+internal actual external val awaitSkiko: Promise<JsAny>
