+++ skiko/src/webMain/kotlin/org/jetbrains/skia/webext/ArrayBuffer.web.kt
@@ -6,27 +6,26 @@ import org.jetbrains.skia.impl.Native
 import org.jetbrains.skia.impl.NativePointer
 import org.jetbrains.skia.impl._malloc
 import org.jetbrains.skiko.ExperimentalSkikoApi
+import org.jetbrains.skiko.await
+import org.jetbrains.skiko.wasm.awaitSkiko
 import org.khronos.webgl.ArrayBuffer
 import org.khronos.webgl.Uint8Array
+import kotlin.js.JsAny
 import kotlin.js.js
 
-internal fun skikoArrayBuffer(skikoWasm: SkikoWasm): ArrayBuffer =
+internal fun skikoArrayBuffer(skikoWasm: JsAny): ArrayBuffer =
     js("skikoWasm.wasmExports.memory.buffer")
-
-internal external interface SkikoWasm
 
 internal suspend fun copyBufferToSkiko(
     src: ArrayBuffer
 ): NativePointer {
     val ptr = _malloc(src.byteLength)
     if (ptr != Native.NullPointer) {
-        val skikoArrayBuffer = skikoArrayBuffer(getSkikoWasm())
+        val skikoArrayBuffer = skikoArrayBuffer(awaitSkiko.await())
         copyBuffer(src, skikoArrayBuffer, ptr)
     }
     return ptr
 }
-
-internal expect suspend fun getSkikoWasm(): SkikoWasm
 
 internal fun copyBuffer(
     src: ArrayBuffer,
